libvb/Constellation: separate declaration/definition, two explicit implementations

+++ tests/test_hypermap.cpp
@@ -11,13 +11,13 @@ int main (int argc, char ** argv) {
 	auto M = HLib().at(H['g']);
 	cout << M << endl;
 
-	Constellation<long double> C {M,H,H['n']};
+	Constellation<double> C {M,H,H['n']};
 	C.findn();
 	if (H['v']) { C.show(); C.belyi(); C.slow=true; C.img->update(); C.img->pause(); delete C.img; }
 	if (!H['q']) { C.belyi(); cout << endl << C; return 0; }
 
 	#ifdef HAVE_COMPLEX_GMP
-	Constellation<gmp100>	Cq = cconvert <long double,gmp100> (C); Cq.findn(); Cq.belyi(); cout << endl << Cq;
+	Constellation<gmp100>	Cq = cconvert <double,gmp100> (C); Cq.findn(); Cq.belyi(); cout << endl << Cq;
 	#else
 	cerr << "No std::complex<boost::numeric::whatever> support!" << endl;
 	#endif
